Move IsCloudProviderEnabled from 'setup' to 'utils'
