+++ pkg/util/ec2/ec2_account_id.go
@@ -10,6 +10,7 @@ import (
 	"fmt"
 
 	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
+	configutils "github.com/DataDog/datadog-agent/pkg/config/utils"
 	"github.com/DataDog/datadog-agent/pkg/util/cachedfetch"
 	ec2internal "github.com/DataDog/datadog-agent/pkg/util/ec2/internal"
 )
@@ -17,7 +18,7 @@ import (
 var accountIDFetcher = cachedfetch.Fetcher{
 	Name: "AWS Account ID",
 	Attempt: func(ctx context.Context) (interface{}, error) {
-		if !pkgconfigsetup.IsCloudProviderEnabled(CloudProviderName, pkgconfigsetup.Datadog()) {
+		if !configutils.IsCloudProviderEnabled(CloudProviderName, pkgconfigsetup.Datadog()) {
 			return "", fmt.Errorf("cloud provider is disabled by configuration")
 		}
